requirements.txt
ipykernel==6.29.5
google-generativeai==0.8.4
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
tqdm==4.67.1
numpy==2.2.2
spacy==3.8.4
jupyter==1.1.1
